Tehtävät 4.18-4.23 tehty. Olipas haastavat vikat tehtävät

+++ backend_blog/controllers/blogs.js
@@ -1,6 +1,8 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -8,10 +10,12 @@ blogsRouter.get('/', async (request, response) => {
       response.json(blogs)
       })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   const body = request.body
-  
-  const user = await User.findById(body.userId)
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'missing or invalid token' }) // 🔹 Muista 401!
+  }
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -30,9 +34,24 @@ blogsRouter.post('/', async (request, response, next) => {
   }      
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: "blog not found" });
+  }
+  if (!user) {
+    return response.status(401).json({ error: "user not found" });
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({ error: "not allowed, wrong user" });
+  }
+  
+  if ( blog.user.toString() === user._id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)  
+    response.status(204).end()
+  } 
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
